# use cache `Set()` for performance improvement